requirements.txt
asgiref==3.8.1
Django==5.0.3
django-admin-thumbnails==0.2.8
django-ckeditor-5==0.2.12
django-debug-toolbar==4.3.0
django-filter==24.2
django-taggit==5.0.1
pillow==10.2.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
